+++ Demo/Application/Features/SEPA Direct Debit/BraintreeDemoSEPADirectDebitViewController.swift
@@ -53,14 +53,14 @@ class BraintreeDemoSEPADirectDebitViewController: BraintreeDemoBaseViewControlle
         billingAddress.region = "Annaberg-buchholz"
         billingAddress.postalCode = "09456"
         billingAddress.countryCodeAlpha2 = "FR"
-        
+
         let sepaDirectDebitRequest = BTSEPADirectDebitRequest()
         sepaDirectDebitRequest.accountHolderName = "John Doe"
-        sepaDirectDebitRequest.iban = generateRandomIBAN()
+        sepaDirectDebitRequest.iban = "FR7618106000321234566666608"
         sepaDirectDebitRequest.customerID = generateRandomCustomerID()
         sepaDirectDebitRequest.mandateType = .oneOff
         sepaDirectDebitRequest.billingAddress = billingAddress
-        sepaDirectDebitRequest.merchantAccountID = "eur_pwpp_multi_account_merchant_account"
+        sepaDirectDebitRequest.merchantAccountID = "EUR-sepa-direct-debit"
         
         if #available(iOS 13.0, *) {
             sepaDirectDebitClient.tokenize(request: sepaDirectDebitRequest, context: self) { sepaDirectDebitNonce, error in
@@ -87,15 +87,6 @@ class BraintreeDemoSEPADirectDebitViewController: BraintreeDemoBaseViewControlle
     
     private func generateRandomCustomerID() -> String {
         String(UUID().uuidString.replacingOccurrences(of: "-", with: "").prefix(20))
-    }
-    
-    private func generateRandomIBAN() -> String {
-        let length = 25
-        let characters = "0123456789"
-        let randomCharacters = (0..<length).map{ _ in characters.randomElement()! }
-        let randomString = String(randomCharacters)
-
-        return "FR" + randomString
     }
 }