+++ Sources/BraintreeSEPADirectDebit/BTSEPADirectDebitNonce.swift
@@ -24,7 +24,7 @@ import BraintreeCore
         guard let nonce = json["nonce"].asString() else { return nil }
         
         self.nonce = nonce
-        self.ibanLastFour = json["details"]["ibanLastChars"].asString()
+        self.ibanLastFour = json["details"]["last4"].asString()
         self.customerID = json["details"]["customerId"].asString()
         self.mandateType = BTSEPADirectDebitMandateType.getMandateType(from: json["details"]["mandateType"].asString())
     }