+++ Sources/BraintreeSEPADirectDebit/SEPADirectDebitAPI.swift
@@ -5,81 +5,84 @@ import BraintreeCore
 
 class SEPADirectDebitAPI {
     
-    // TODO: Use BTAPIClient when in Sandbox
+    private let apiClient: BTAPIClient
+    
+    @objc(initWithAPIClient:)
+    init(apiClient: BTAPIClient) {
+        self.apiClient = apiClient
+    }
 
     func createMandate(
         sepaDirectDebitRequest: BTSEPADirectDebitRequest,
         completion: @escaping (CreateMandateResult?, Error?) -> Void
     ) {
-        guard let sepaDirectDebitData = try? JSONEncoder().encode(sepaDirectDebitRequest) else {
-            completion(nil, SEPADirectDebitError.encodingFailure)
-            return
-        }
+        let json: [String: Any] = [
+            "sepa_debit": [
+                "merchant_or_partner_customer_id": sepaDirectDebitRequest.customerID ?? "",
+                "mandate_type": sepaDirectDebitRequest.mandateType?.description ?? "",
+                "account_holder_name": sepaDirectDebitRequest.accountHolderName ?? "",
+                "iban": sepaDirectDebitRequest.iban ?? "",
+                "billing_address": [
+                    "address_line_1": sepaDirectDebitRequest.billingAddress?.streetAddress,
+                    "address_line_2": sepaDirectDebitRequest.billingAddress?.extendedAddress,
+                    "admin_area_1": sepaDirectDebitRequest.billingAddress?.locality,
+                    "admin_area_2": sepaDirectDebitRequest.billingAddress?.region,
+                    "postal_code": sepaDirectDebitRequest.billingAddress?.postalCode,
+                    "country_code": sepaDirectDebitRequest.billingAddress?.countryCodeAlpha2
+                ]
+            ],
+            "merchant_account_id": sepaDirectDebitRequest.merchantAccountID ?? "",
+            "cancel_url": sepaDirectDebitRequest.cancelURL,
+            "return_url": sepaDirectDebitRequest.returnURL
+        ]
 
-        let request = buildURLRequest(withComponent: "merchants/pwpp_multi_account_merchant/client_api/v1/sepa_debit", httpBody: sepaDirectDebitData)
-        
-        let task = URLSession.shared.dataTask(with: request) { data, response, error in
-            guard let data = data, error == nil else {
-                DispatchQueue.main.async {
-                    completion(nil, error)
-                }
+        apiClient.post("v1/sepa_debit", parameters: json) { body, response, error in
+            self.apiClient.sendAnalyticsEvent("ios.sepa-direct-debit.api-request.create-mandate.started")
+            if let error = error {
+                self.apiClient.sendAnalyticsEvent("ios.sepa-direct-debit.api-request.create-mandate.error")
+                completion(nil, error)
                 return
             }
-            
-            do {
-                let result = try JSONDecoder().decode(CreateMandateResult.self, from: data)
-                DispatchQueue.main.async {
-                    completion(result, nil)
-                }
-            } catch {
-                completion(nil, SEPADirectDebitError.invalidResult)
+
+            guard let body = body else {
+                self.apiClient.sendAnalyticsEvent("ios.sepa-direct-debit.api-request.create-mandate.no-body.error")
+                completion(nil, SEPADirectDebitError.noBodyReturned)
+                return
             }
+
+            self.apiClient.sendAnalyticsEvent("ios.sepa-direct-debit.api-request.create-mandate.success")
+            let result = CreateMandateResult(json: body)
+            completion(result, nil)
         }
-        task.resume()
     }
-    
+
     func tokenize(createMandateResult: CreateMandateResult, completion: @escaping (BTSEPADirectDebitNonce?, Error?) -> Void) {
         let json: [String: Any] = [
             "sepa_debit_account": [
-                "iban_last_chars": createMandateResult.ibanLastFour,
-                "customer_id": createMandateResult.customerID,
+                "last_4": createMandateResult.ibanLastFour,
+                "merchant_or_partner_customer_id": createMandateResult.customerID,
                 "bank_reference_token": createMandateResult.bankReferenceToken,
                 "mandate_type": createMandateResult.mandateType
             ]
         ]
-        
-        guard let jsonData = try? JSONSerialization.data(withJSONObject: json) else {
-            completion(nil, SEPADirectDebitError.jsonSerializationFailure)
-            return
-        }
 
-        let request = buildURLRequest(withComponent: "merchants/pwpp_multi_account_merchant/client_api/v1/payment_methods/sepa_debit_accounts", httpBody: jsonData)
-        
-        let task = URLSession.shared.dataTask(with: request) { data, response, error in
-            guard let data = data, error == nil else {
-                DispatchQueue.main.async {
-                    completion(nil, error)
-                }
+        apiClient.post("v1/payment_methods/sepa_debit_accounts", parameters: json) { body, response, error in
+            self.apiClient.sendAnalyticsEvent("ios.sepa-direct-debit.api-request.tokenize.started")
+            if let error = error {
+                self.apiClient.sendAnalyticsEvent("ios.sepa-direct-debit.api-request.tokenize.error")
+                completion(nil, error)
                 return
             }
 
-            let result = BTSEPADirectDebitNonce(json: BTJSON(data: data))
-            DispatchQueue.main.async {
-                completion(result, nil)
+            guard let body = body else {
+                self.apiClient.sendAnalyticsEvent("ios.sepa-direct-debit.api-request.tokenize.no-body.error")
+                completion(nil, SEPADirectDebitError.noBodyReturned)
+                return
             }
+
+            self.apiClient.sendAnalyticsEvent("ios.sepa-direct-debit.api-request.tokenize.success")
+            let result = BTSEPADirectDebitNonce(json: body)
+            completion(result, nil)
         }
-        task.resume()
-    }
-    
-    func buildURLRequest(withComponent component: String, httpBody: Data) -> URLRequest {
-        let baseURL = URL(string: "http://localhost:3000")?.appendingPathComponent(component)
-
-        var request = URLRequest(url: (baseURL)!)
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.setValue("development_testing_pwpp_multi_account_merchant", forHTTPHeaderField: "Client-Key")
-        request.httpMethod = "POST"
-        request.httpBody = httpBody
-
-        return request
     }
 }