+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Braintree iOS SDK Release Notes
 
+## unreleased
+* BraintreeSEPADirectDebit
+  * Add support for SEPA Direct Debit for approved merchants through the Braintree SDK
+  * SEPA Direct Debit is only available to select merchants, please contact your Customer Support Manager or Sales to start processing SEPA bank payments
+  * Merchants should use the `BTSepaDirectDebitClient.tokenize` method while passing in the `BTSEPADirectDebitRequest` and `context` while conforming to `ASWebAuthenticationPresentationContextProviding`
+
 ## 5.10.0 (2022-06-06)
 * Fix potential crash when http request fails with no error but empty data (thanks @cltnschlosser)
 * Update Cardinal SDK to version 2.2.5-3